AIAgentWithLG/requirements.txt
openai
langchain
python-dotenv
langchain-community
langchain-core
